Windows warning fixes.
svn path=/trunk/libavg/; revision=3232

+++ src/player/CurveNode.h
@@ -83,7 +83,7 @@ class CurveNode : public VectorNode
                 double opacity, bool bParentDrawNeeded);
 
     private:
-        double getCurveLen();
+        int getCurveLen();
         void updateLines();
         void addLRCurvePoint(const DPoint& pos, const DPoint& delta);
         DPoint m_P1;
